Move requests to index.requests.js to break up index.js
